make smaller components
