+++ src/styles/StorageIndicator.css
@@ -1,0 +1,38 @@
+.StorageUsed {
+  position: relative;
+  font-variant-numeric: tabular-nums;
+  height: fit-content;
+  background-color: var(--green-1);
+  border: 1px solid oklch(50.1% 0.1 143.9 / 0.5);
+  color: oklch(91.2% 0.093 144.8 / 0.925);
+  box-shadow: oklch(67.2% 0.11 146 / 0.1) 0px -4px 9px 3px inset;
+  cursor: help;
+  padding: 0.15rem 0.5rem 0.15rem 1.25rem;
+  border-radius: 99px;
+  gap: 0.25rem;
+  font-weight: 300;
+
+  transition: background-color, box-shadow 200ms ease-out;
+}
+
+.StorageUsed:hover {
+  /* background-color: rgba(27, 65, 27, 0.87); */
+  box-shadow: oklch(67.2% 0.11 146 / 0.4) 0px -4px 9px 3px inset;
+  .CircleBackground {
+    background-color: oklch(74.8% 0.227 142.9);
+  }
+}
+
+.CircleBackground {
+  position: absolute;
+  left: 0.4rem;
+  top: 0.5rem;
+  background-color: oklch(82.9% 0.192 143.6 / 0.87);
+  border-radius: 99px;
+  width: 0.6rem;
+  height: 0.6rem;
+  filter: blur(4px);
+  -webkit-filter: blur(4px);
+
+  transition: transform 200ms ease-in-out;
+}
