+++ src/styles/StrengthCircle.css
@@ -1,0 +1,5 @@
+.StrengthCircle {
+  position: relative;
+  z-index: 10;
+  cursor: help;
+}
